Don't zombie on preexisting Haste collisions on startup
Summary:
For historical reasons, Metro currently fails "hard" (actually, tries to, but enters a zombie state) when the filesystem contains Haste collisions on startup.

This is mostly inherited from Jest, which prefers to fail fast on the Haste (or Jest Mock) map being in a bad state.

In Metro, it isn't useful to anyone - the server is designed to gracefully handle temporary Haste collisions - warn about them, fail to resolve ambiguous Haste names with an actionable message, and recover when the collision is fixed.

This diff changes the argument to `metro-file-map` to prevent a throw which would effectively hang Metro, allowing it to treat Haste collisions detected on startup exactly the same as a Haste collision created while Metro is running.

Changelog:
```
 - **[Fix]**: Don't enter zombie state on startup Haste collisions
```

Reviewed By: huntie

Differential Revision: D66935485

fbshipit-source-id: 39c664a12bb5be8c12851cab49dec66213daa95a