Simplify symlink handling in metro-file-map processFile (#1398)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1398

Reading symlink targets with `readLink` is currently delegated to `metro-file-map`'s worker abstraction. However, due to the way the abstraction is used for symlinks, we never actually use the worker processes/threads, and always process in band.

The key here is that `readLink` is always mutually exclusive with other worker tasks, like `computeSha1` and `computeDependencies` - as it must be, because those operations don't make sense for symlink files.

This lifts the call to `readLink` out of the abstraction and explicitly into the main process, simplifying the worker and improving readability.

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D66899343

fbshipit-source-id: 75e40a4cb7721183fdbcc959c96933e08d175659