Flow types: strengthen change event guarantee, fix missing type coverage
Summary:
`metadata` on file change events out of metro-file-map is typed as optional and nullable, but in fact is always set within metro-file-map, and DeltaCalculator would throw if it wasn’t - Flow didn’t pick that up only because a $FlowFixMe hid the problem.

This is a Flow-only change to reflect that and improve coverage.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D67259536

fbshipit-source-id: 7e21d6d948e05a4cfd69eb38f898940075ebab2e