FSEventsWatcher: use common recReadDir
Summary:
`FSEventsWatcher` has its own implementation of `recReaddir` which is almost identical to the one in `./common.js`, except that it wraps handlers in `path.normalize`.

There’s only one handler, so we can trivially remove this proxy abstraction and reuse the common `recReaddir`.

(The end goal here is to slim down and tidy up FSEventsWatcher ahead of using it as a basis for native recursive watching)

Changelog: Internal

Differential Revision: D67286092