FSEventsWatcher: use common recReadDir (#1409)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1409

`FSEventsWatcher` has its own implementation of `recReaddir` which is almost identical to the one in `./common.js`, except that it wraps handlers in `path.normalize`.

There’s only one handler, so we can trivially remove this proxy abstraction and reuse the common `recReaddir`.

(The end goal here is to slim down and tidy up FSEventsWatcher ahead of using it as a basis for native recursive watching)

Changelog: Internal

Reviewed By: huntie

Differential Revision: D67286092

fbshipit-source-id: 8a72f5dead13b0eae25e03bcc998397332ed58e6