FSEventsWatcher: Don't stat ignored files
Summary:
The include/exclude logic in `metro-file-map` is a bit of a mess, but roughly it's meant to work as follows:

 - Paths matching `ignorePattern`, a `?RegExp`, are not crawled (added to `TreeFS`) and do not result in emitted events. Note that `ignorePattern` may match directories, in order to ignore anything inside them. This is important to prevent crawling whole subtrees unnecessarily.
 - *Regular files* must match the given `globs` (`Array<string>`). Directories and symlinks are *not* filtered by glob. Metro uses `globs` to limit the file extensions crawled and watched, eg `**/*.js`.

(Note that there are all sorts of caveats here - it's not consistent whether we match whole paths, project-relative paths, or watch-root relative paths, nor whether we normalise to posix separators before matching. This needs properly unpicking.. some other time.)

Because `globs` are only applied to regular files, we must `lstat` a changed path to know whether we should filter it with `globs`. However, `ignorePattern` sometimes allows us to filter out a path early, before `lstat`.

By splitting out `isIncluded` from `doIgnore` (and tweaking the naming a bit for clarity), we can avoid some unnecessary I/O.

Changelog:
```
 - **[Performance]**: Don't stat ignored files in fsevents watcher.
```

Differential Revision: D67593664