metro-file-map: Refactor out retainAllFiles mutation
Summary:
Refactor to clarify the roles of `ignore` and `ignorePattern` in the crawl and watch stages of `metro-file-map` startup, and remove the unnecessary mutation of `retainAllFiles`.

By removing this read of `retainAllFiles` from `_processFile`, we can more easily extract/abstract file processing, which will be necessary for the addition of a buildozer-powered Buck file plugin for `metro-file-map`.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D67652663

fbshipit-source-id: 46f2378f73d30fdffde7bb39fe10ee0885af940f