Watcher backends: refactor common functionality into base watcher, improve type safety
Summary:
Refactoring of Watcher backends to improve type safety and reduce duplication, by moving some functionality into a base class, and making private the `any`-typed `EventEmitter` in favour of strongly typed interface/base methods.

This also introduces the `startWatching(): Promise<void>` API to backends, replacing the old mechanism of starting the process in the constructor and emitting `'ready'` when done - this was awkward for watchers that started synchronously, as noted in code comments deleted here.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D67622921

fbshipit-source-id: 2eb2d789ab63f1c311a5028ea252053faf32ca4f