Refactor: NodeWatcher->FallbackWatcher, encapsulate fallback walk
Summary:
Rename "NodeWatcher" largely to avoid confusion with the upcoming `NativeWatcher`, which also uses Node.

"`FallbackWatcher`" is chosen because by design it uses the most basic, always-available APIs, as opposed to `WatchmanWatcher` or `FSEventsWatcher` that have optional dependencies, but are preferred where available.

Also moves `recReaddir` from `common` into `FallbackWatcher` as it's now only used there.

(Note that as this is part of `metro-file-map`, the API is semver unstable, and safe to change)

Changelog: Internal

Reviewed By: huntie

Differential Revision: D67634394

fbshipit-source-id: 13560c3742fa32a9ca7f596afc872a7bf18e7e76