Tests: fix inconsistent use of configured vs emitted root
Summary:
This test currently mocks watcher backends by:
 1. Using a different emitter instance for each watched root
 2. calling `emit('all', payload)`, where `payload` includes `root`.

In reality, these roots should never differ, but in practice appending to the `root` in the payload is equivalent to prepending to the (root-)`relativePath`.

Therefore some of these tests currently pass, but the mock behaviour isn't realistic.

This amends the tests to always emit the configured `root` in the payload, and corrects `relativePath`s accordingly.

(Helpful for the next diff, where the amended behaviour is enforced by a tighter API)

Changelog: Internal

Reviewed By: huntie

Differential Revision: D67636697

fbshipit-source-id: 7d524348dde378fc982fe517b179a4ae8b171b7c