Fix watcher path normalisation with Windows and Watchman (#1423)
Summary:

Fixes an issue with the Watchman watcher backend on Windows where emitted posix-style paths were not properly normalised, leading to inconsistencies with other watchers and broken tests.

Adds the necessary normalisation and reenables tests.

Changelog: Internal

(This is luckily corrected downstream by the use of `path.join` on `relativePath`, which performs normalisation, so there was no observable bug here, it was just fragile.)

Reviewed By: huntie

Differential Revision: D67700720