Fix watcher path normalisation with Windows and Watchman (#1423)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1423

Fixes an issue with the Watchman watcher backend on Windows where emitted posix-style paths were not properly normalised, leading to inconsistencies with other watchers and broken tests.

Adds the necessary normalisation and reenables tests.

Changelog: Internal

(This is luckily corrected downstream by the use of `path.join` on `relativePath`, which performs normalisation, so there was no observable bug here, it was just fragile.)

Reviewed By: huntie

Differential Revision: D67700720

fbshipit-source-id: 08ae84c0473e9db1b4e1844210ba111fe25fbc92