Add unstable_inlinePlatform option to transform config
Summary:
Metro resolver logic can be customized in various ways to support bespoke platform-style extensions in modules. In certain cases, these kinds of bespoke platform overrides can switch Platform.select behaviors. However, when used in conjunction with the default inlinePlatform transform for Metro, this kind of override may not be possible.

This change introduces an unstable API that can be used to disable inlinePlatform transforms.

Reviewed By: robhogan

Differential Revision: D68346407

fbshipit-source-id: d3042554b21b437e85ced7adf5156ecad81263d5