Refactor: allow options other than buildParameters to be passed to CacheManagers
Summary:
Internal refactor of `CacheManager` / `CacheManagerFactory` to allow an object with a `buildParameters` prop, rather than passing `buildParameters` directly as an argument.

`BuildParameters` is strictly defined as parameters that may change the build output, and every prop is enforced as a member of the cache key. This change creates room for other input passed from `metro-file-map`'s constructor to the given factory function, like perfLogger, callbacks, etc. We'll use it in subsequent diffs to implement reactive caches (auto save on changes, etc).

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D68090141

fbshipit-source-id: a8419a592dbbe3a49595c2b60dbf32b5616f7095