Refactor: separate Haste and mocks updates from worker processing
Summary:
Refactors the way the Haste and mocks maps are updated during initial load from cache, after a crawl to update the initial state, and on changes during watch mode.

This makes mutations of `MutableHasteMap`, etc, synchronous with mutations of `TreeFS`, so that they're guaranteed to be consistent. Previously `MutableHasteMap` was updated with an async function (though in practice, right at the end of it), and `TreeFS` only after that promise resolved.

Potentially, this allows us to persist state at any point, and be certain that it's internally consistent. That's going to be useful for efficient lazy hashing...

Changelog: Internal

Reviewed By: huntie

Differential Revision: D67793946

fbshipit-source-id: 0a0f9a3dadf2882c428d0118fab559fc1997a1ac