WatchmanWatcher: update clock on events
Summary:
We have three pieces of state making up the file map cache:

 - `fileSystem` (a tree of files and metadata)
 - `plugins` (Haste/mock data, eg Haste names to paths)
 - `clocks` (the Watchman-only file clocks associated with the current state)

These are all updated during the initial crawl, and saved in the file map cache. `clocks` is used to perform `since` queries with Watchman on the next startup.

Currently though, when we observe changes during **watch mode**, we update `fileSystem` and `plugins` (synchronously with each other) but never update `clocks`, because we currently have no use for `clocks` after saving the cache, which we only do at the end of startup.

This uses the `clock` we receive from Watchman on every subscription response to also keep `clocks` perfectly in sync with `fileSystem` and `plugins`.

## Why?
By ensuring `clocks` is synchronously up-to-date with `fileSystem` and `plugins` state, we can, at any point we choose, take a correct snapshot of all cacheable state and persist it.

This allows us (in later diffs) to build an initial cheap state, lazily populate it (eg, by building a bundle, and hashing only the files needed for the bundle), and save it once populated. This gives us the best of both worlds - fast cold startup, without regressing build performance on warm starts.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D67825634

fbshipit-source-id: c387c5e572ad29929bcfa08c4e2bbf94250f1dc3