Cache auto-save 3/3: Expose watcher.unstable_autoSaveCache, default enabled (#1434)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1434

Configures a new mechanism for the file map cache to re-save after changes made to the file system while Metro is running.

Currently, If you make a change while Metro is running, Metro will re-process (hash) that file and update an in-memory record, but won't save it. The next time Metro starts up, it will process that file again, and write the cache at the end of startup.

In most cases this provides a modest reduction in startup work/time - proportional to the number of changes in the previous Metro session.

However, it unlocks the potential of doing more processing lazily - ie, we don't need to hash everything up front if we can hash only what we need for a bundle, and then save those hashes in a cache.

We'll use this to implement lazy hashing.

## Implementation
The auto save mechanism is contained within `DiskCacheManager` via generic listening APIs. When the file map emits a `change` event to Metro, we start/restart a configurable debounce timer, and save the cache asynchronously.

The 5 second default is chosen so that we don't try to save while Metro may be busy - eg, a change is likely to be followed by an HMR exchange over the next second or two.

Changelog:
```
 - **[Experimental]**: Auto-save file cache via `config.watcher.unstable_autoSaveCache`
```

Reviewed By: huntie

Differential Revision: D69024198

fbshipit-source-id: 1ab06fba567e981bfa59f0e2f045ee3bcb75e3d5