metro-file-map: Allow FileProcessor to return content (lazy hashing 1/3)
Summary:
## Stack
In this stack we're moving towards metro-file-map being able to *lazily* compute file metadata - in particular the SHA1 hash - only when required by the transformer.

More context in https://github.com/facebook/metro/issues/1325#issuecomment-2303019051

## Reading files at most once
With the exception of Eden+Watchman, computing the SHA1 requires reading the content, but if we lazily compute the SHA1, find no cache hit (likely, given no cached SHA1 implies no local transform cache), we'll then want to transform the file we've just read.

This change will allow us to pass the content we read to derive the SHA1 through to the transform worker *without reading the file from disk a second time*.

`maybeReturnContent` is admittedly a slightly awkward API, but it's Metro-internal, and it was the best I could come up with for returning content only if it's already been read (otherwise, it's faster for transform workers to read it from disk), and only if needed (otherwise there's potentially significant extra IPC for no reason).

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D69372887