metro-file-map: Spawn fewer workers for smaller workloads
Summary:
`metro-file-map` currently spins up `maxWorkers` workers (processes, or threads if `enableWorkerThreads`) for any batch workload. `maxWorkers` is `os.availableParallelism` by default.

When we have a warm cache or Saved State, we may only have a handful of changed files to process, and frequently create more workers than we have files.

This implements a simple heuristic factor `maxFilesPerWorker`, such that we use `Math.ceil(numFiles/maxFilesPerWorker)` parallelism (where 1 "worker" means we process in-band).

The default value of  100 is somewhat finger-in-air as a bunch of factors would determine the optimal value - including threads or not, which host platform, and computational cost of the haste impl and dependency extraction. We could expose this as Metro configuration in future.

Changelog:
 - **[Performance]** Don't start an excessive number of workers for hashing files during startup.

Reviewed By: huntie

Differential Revision: D69704168

fbshipit-source-id: 68470470dd85cbc20f94cd5fa6f88ad11c090c01