Add package.json "exports" to Metro packages, alias "private" to "src" (#1441)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1441

Add an `exports` field to the `package.json` of each Metro package, exposing all current exports, and a new `./private/x` alias for `./src/x.js`.

This is intended to be non-breaking - all current uses of deep imports to any source file in Metro should continue to work (whether or not they use extensionless or directory imports).

We'll encourage any consumers of `metro*/src/` imports to either move to `metro*/private/` with no semver guarantees - or accommodate their use case with a public API where it makes sense.

The high-level idea here is to keep Metro "hackable" for integrators and not immediately break anyone, but to allow us to move faster by reducing the API surface that's implicitly semver-stable, but in most cases has no public consumers.

Changelog:
```
 - **[Deprecated]**: Deprecate deep `src` imports e.g. `metro/src/x.js` across all `metro*` packages, pending removal from the stable API.
 - **[Experimental]**: Expose deep imports of Metro packages (`metro/src/x.js` as `metro/private/x`) as explicitly private (non-semver).
```

Reviewed By: huntie

Differential Revision: D70038572

fbshipit-source-id: 6e0cbaca052cee8d15d961c7ce0035797e244c08