Fix global/Haste package resolution with lazy SHA1
Summary:
`metro-file-map` skips "processing" of files (ie, passing the file paths to `FileProcessor` and potentially a worker) when there is "nothing to do" - this helps appropriately scale the worker pool.

I missed one possibility though - with `enableHastePackages`, the worker must visit each `package.json` outside `node_modules` to read its `package.json#name`, which becomes the ID of the global/Haste package. This is true even if `hasteImplModulePath == null`, ie Haste isn't being used for modules.

This amends the file-skipping logic not to skip if `enableHastePackages` is true and the file might be a `package.json`.

Changelog:
```
 - **[Experimental]**: Fix global package resolution with `unstable_lazySha1`
```

Reviewed By: hoxyq

Differential Revision: D70473017

fbshipit-source-id: c3472314d979bdd5da1c7604e3349e40d725bad2