Breaking: Enable lazy hashing by default, changes to Transformer and DependencyGraph APIs (#1451)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1451

Lazy hashing was released in Metro 0.81.2 as an opt-in. It significantly reduces the IO Metro needs to perform on startup and generally improves cold start speed >3x, while being fully compatible with cache lookups even on a cold start.

 {F1975587641}

(We announced this in https://x.com/MetroBundler/status/1895102434889850960 )

There's very little risk to this feature - the technically-breaking change required only affects integrators using an "internal" API - instantiating a `Transformer`.

This flips it to default-on ahead of Metro 0.82.0

Changelog:
```
 - **[Breaking]**: Integrators consuming `metro/src/DeltaBundler/Transformer` must now pass an options object with `getOrComputeSha1`
 - **[Breaking]**: `getSha1` removed from `metro/src/node-haste/DependencyGraph`, prefer `getOrComputeSha1`
 - **[Performance]**: Enable lazy file hashing by default (disable with `watcher.unstable_lazySha1: false`)
```

Reviewed By: huntie

Differential Revision: D70463022

fbshipit-source-id: bf0fec70eaab4960142f260703a349d422cecf0a