fix terminal status log updates
Summary:
Since Node v12.7.0, `readline` functions are [using a callback](https://nodejs.org/api/readline.html#readlineclearlinestream-dir-callback) and are not guaranteed to finish execution until these callbacks are called.
`clearStringBackwards` in particular, was calling readline functions many times over, which was creating issues with how we were clearing the status, but the rest of the Terminal was also using async functions as if they were sync.

Changelog:
[General][Fix] fix terminal statuses incorrectly persisting wrong lines

Reviewed By: robhogan

Differential Revision: D70779455

fbshipit-source-id: 77f48acc3e0aa448b907d3b24db8b24c795a581f