Fix edge-case overmatching of pattern keys in package.json exports/imports resolution
Summary:
Currently package exports resolution pattern matching may falsely match some patterns because we don't apply a subtle part of the [specification](https://nodejs.org/docs/latest/api/esm.html#resolution-algorithm-specification) - specifically, we check that the subpath starts with the pattern base and ends with the pattern trailer, but not that the subpath length >= the pattern length.

This means the `startsWith` and `endsWith` checks may both be satisfied by the same substring, e.g:

```
  "exports": {
    "./repeated*/repeated": null
  }
```

Will incorrectly match the substring `./repeated`, because it both starts with `./repeated` and ends with `/repeated`.

The spec would disallow this because `matchKey` is shorter than `patternKey`:

> If `patternTrailer` has zero length, or if `matchKey` ends with `patternTrailer` and the length of `matchKey` is greater than or equal to the length of `expansionKey`, then:
Let `target` be the value of `matchObj[expansionKey]`.
...

This diff applies that bit of the spec.

```
 - **[Fix]**: Fix edge-case overmatching of pattern keys in package.json exports/imports resolution.
```

Reviewed By: huntie

Differential Revision: D70971859

fbshipit-source-id: d3a3c317234bd015b58b7ac784aa5ac8f7c48d5e