Fix handling of export map targets with zero or 2+ wildcard characters in package.json imports/exports resolution
Summary:
Currently, metro-resolver will break if an exports map includes a pattern key whose target does not have exactly one `*` character.

This is an incorrect limitation according to the spec - values may have any number (including zero) `*` characters (*keys* must have no more than one). The spec requires that *all* occurrences of `*` in the matched target are replaced with the `patternMatch` substring of the specified subpath.

Worse, we *`break`* rather than `continue` on such an "invalid" target, which means the presence of one of these targets can mean we fail resolution before other keys in the export map are tried.

This diff implements the spec:

[https://nodejs.org/docs/latest/api/esm.html#resolution-algorithm-specification](https://nodejs.org/docs/latest/api/esm.html#resolution-algorithm-specification)

> If `patternMatch` is a `String`, then
> Return **`PACKAGE_RESOLVE`**(`target` with every instance of "*" replaced by `patternMatch`, `packageURL + "/"`).

```javascript
 - **[Fix]**: Fix handling of export map targets with zero or 2+ wildcard characters in package.json imports/exports resolution
```

Reviewed By: huntie

Differential Revision: D70973129

fbshipit-source-id: 68632b68dbc37dc24988cd68529a5cebfcb5a5d3