feat(metro): Improve stack frame handling by showing the first valid frame when all frames are collapsed (#1483)
Summary:
When an error comes from the internals of your app it can be especially devastating. This generally happens when you have a versioning issue from upgrading. The stacks are currently optimized for user-space errors which make up the majority of issues, but when all frames are collapsed we should assume the issue is an internal bug and show anything to give the user some sort of hint for what to fix.

<!--
Changelog entries should be prefixed with one of the following scopes:
[Breaking, Feature, Fix, Performance, Deprecated, Experimental, Internal]

Examples:
  Changelog: [Fix] Respond with HTTP 404 when a bundle entry point cannot be resolved
  Changelog: [Internal]
-->

Changelog: [Feature] Fallback to showing first collapsed frame when all frames in a stack are collapsed.

Pull Request resolved: https://github.com/facebook/metro/pull/1483

Reviewed By: huntie

Differential Revision: D72973523

Pulled By: robhogan

fbshipit-source-id: 9c8c0585c123d9e3da27fe2745e8e7f8f4eebf53