Refactor out unused `bundleType` from BundleOptions
Summary:
`bundleType` is only used by Metro's reporter (to visually distinguish `BUNDLE` from `MAP`, etc) upon receipt of a request to Metro server, and is unnecessarily part of `BundleOptions` as passed to serialisers.

Within the server, it's redundant because we've already inspected the `pathname` extension of `.bundle`, etc, before deciding which request processor to use - it's only that that is load bearing on the output - so we can simply make it part of creating the request processor, instead of passing the same string to every call to the processor.

This diff removes `bundleType` from the `BundleOptions` type while retaining it in the return type of `parseOptionsFromUrl` (a public API) for now.

Callers may continue to pass it - it'll be unused as it always has been - so this is non-breaking.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D74150833

fbshipit-source-id: 6b0b46b0819f77259c1a60b6a0c09418991be6a9