runBuild: Add ability to return assets
Summary:
Currently, integrators orchestrating a full app build with Metro must either use custom serializers or create a `Server` instance on which to call `getAssets`, because the otherwise convenient `runBuild` API does not call or expose `Server.getAssets`.

Eg, the `community-cli` `bundle` command almost entirely reimplements `runBuild` so that it can obtain an asset listing:

https://github.com/facebook/react-native/blob/9c5fc7958183963f1bed54fc4c1a68983bf178f3/packages/community-cli-plugin/src/commands/bundle/buildBundle.js#L128

This uses APIs that will shortly be private in Metro.

By adding the `assets` boolean to `RunBuildOptions`, we can straightforwardly return an asset list and migrate `bundle` to this public API.

Changelog:
```
 - **[Feature]**: Support returning a list of `AssetData` from `Metro.runBuild`

Reviewed By: huntie

Differential Revision: D74148552

fbshipit-source-id: b0b3052699a8efc6f601dc8080a8ad8df6ed8c67