Export additional Flow types from `metro` entry file (`AssetData`, `Reporter`, etc)
Summary:
These Flow types are used legitimately by integrators including `react-native/community-cli-plugin` and should be part of Metro's public API, which means exposing them through either the package entry point or a public subpath.

This re-exports them through the entry file. The types in this diff along with the refactoring in D74151840 allow us to eliminate all uses of to-be-private Metro APIs in RN itself.

Changelog: Internal (types)

Reviewed By: huntie

Differential Revision: D74147394

fbshipit-source-id: 4144152ffbc00c9e7ba75da229797f5bd556eaea