Constant folding: Don't fold non-literal `void` expressions (#1506)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1506

Fixes: https://github.com/facebook/metro/issues/1505

`void` expressions always evaluate to `undefined` but would rarely be used to express a constant (with the exception of `void 0`, which is a common shortening of `undefined`).

More often, they're used to discard the value of an effectful expression, eg `() => void foo()`, or even `() => void obj.prop` where property access has side effects.

Interpret these more conservatively as not foldable unless we can be confident that they have no side-effects because the argument is a literal.

Changelog:
```
 - **[Fix]**: Don't constant-fold potentially side-effectful void expressions
```

Reviewed By: huntie

Differential Revision: D74403187

fbshipit-source-id: e8e4bd6476f761c3b2f7d4f95e4157d890bad116