Fix optional dependencies shifting dependency indices, allow nulls in dependency arrays (#1522)
Summary:
In Metro, optional dependencies are requires or async imports whose failure to resolve is not treated as a build time error. Currently `import()` or `require()` calls statically within a `try {}` are considered optional, when `transformer.allowOptionalDependencies` is enabled.

## The problem

A longstanding bug exists with optional dependencies due to the way they are filtered out if they cannot be resolved during graph traversal, so that the array of resolved dependencies is shorter than the array of collected dependencies.

| idx | collected dependencies | resolved dependencies |
|---|-------------| --
| 0 | './a.js'        | '/project/a.js'
| 1 | './optional.js' | '/project/c.js'
| 2 | './c.js'        |

This breaks the relationship between a module's `_dependencyMap` and the indices used in Metro `require` calls, assigned before resolution, eg:

`require('./optional.js')`

Is transformed to:

`_$$_REQUIRE(_dependencyMap[1], "./optional.js")`

Which will in fact load the module at `/project/c.js`, by the table above. Moreover, `require('./c.js')` will throw at runtime because `_dependencyMap[2]` is `undefined`.

Note that this behaves roughly as expected *only* if all unresolved dependencies are at the end of the array (later in the AST traversal of dependency collection).

## This change

Here, rather than filtering out unresolved dependencies, we retain them, propagating them to `null`s in the `_dependencyMap`, thus preserving ordering.

| idx | collected dependencies | resolved dependencies |
|---|-------------| --
| 0 | './a.js'        | '/project/a.js'
| 1 | './optional.js' | null
| 2 | './c.js'        | '/project/c.js'

We also explicitly handle them in `metroRequire` to throw a runtime "Cannot find module" error, consistent with Node.js.

## Changelog
```
 - **[Fix]** Fix `require` following missing optional dependencies returning wrong modules or failing to load
```

Pull Request resolved: https://github.com/facebook/metro/pull/1522

Test Plan:
## Before
Because unresolvable dependencies are simply omitted from the `_dependencyMap` of the module, dependency indices are off-by-one, with bizarre results:

 {F1979783549}

## After

Module IDs appear at the expected positions in dependency maps, so dependencies indexed after an optional dependency are correct:
{F1979791272}

`metroRequire` specifically handles nulls and throws a runtime error with a message consistent with Node.js

 {F1979791102}

Reviewed By: vzaidman

Differential Revision: D77254860

Pulled By: robhogan

fbshipit-source-id: 4c343f42b2dfc25318770e7f977f476c2cec8826