Stop publishing metro-memory-fs
Summary:
`metro-memory-fs` is a general-purpose in-memory mock Node.js file system for use in testing - it doesn't really have anything to do with Metro.

When it was created, we probably didn't have many alternatives. But `memfs` (for example) is widely used and well maintained.

`metro-memory-fs` is used pretty extensively in Metro's own tests and some Meta-internal tests, so we can't just delete it, but this moves it out of the way of OSS publishing tooling so we don't have to worry about supporting it for OSS use cases.

Folks can of course continue to use the last published version 0.82.5, but won't receive updates.

Changelog:
```
 - **[Removed]**: `metro-memory-fs` will no longer be published
```

Reviewed By: vzaidman

Differential Revision: D78167490

fbshipit-source-id: 7c112c177acb1fee92276f644457903b1c5c6965