File map: introduce worker.setup for unchanging worker args
Summary:
Recent versions of `jest-worker` support exposing a `setup` API that may be used for worker initialisation.

We can take advantage of this in `metro-file-map` to `require` things like the Haste impl, and (shortly) plugin workers as part of initialisation, which saves passing constants like module paths over IPC repeatedly with each workload. (Currently, we send `hasteImplModulePath` - a potentially quite long absolute path - with every workload, and the worker has to check that it doesn't vary).

Note that in some cases we use the worker outside of a `jest-worker` context - e.g. to process a single file in-band, we simply import and call the function. To support `setup`, I've refactored this to a class which is constructed with setup args.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D72854450

fbshipit-source-id: ba222f7d8feb53abcc9c82a55103f38546dddeea