File map: Move Haste ID from index 0 to index 6
Summary:
As part of introducing generalised file map plugins, we'll want to store plugin data in a predictable, consistent manner in the file map cache, rather than having hardcoded indices for each type of data.

Currently, all data corresponding to a file is stored in a flat tuple:

```
[
  /* 0: haste id */ '',
  /* 1: mtime */ null,
  /* 2: size */ 0,
  /* 3: visited */ 1,
  /* 4: dependencies */ '',
  /* 5: sha1 */ null,
  /* 6: symlink */ 0,
],
```

This data structure is critical to file map performance on larger projects. To maintain an optimal format for serialisation, memory, and size on disk, it'd be best to keep the flat tuple, but rearrange it so that an arbitrary number of plugins can store their data in the end elements:

In particular, we have `HastePlugin` and we'd like to store its data in the "plugin data" space. So this diff moves it to the end, in preparation for generalising:

```
[
  /* mtime */ null,
  /* size */ 0,
  /* visited */ 1,
  /* dependencies */ '',
  /* sha1 */ null,
  /* symlink */ 0,
  /* plugin 1: Haste */ '',
  /* more plugins... */
],
```

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D76755138

fbshipit-source-id: 148fe3d5c144b0d1bfe0e7111f7b5e259f46f726