Convert to ESM-syntax exports: metro-source-maps,metro-symbolicate,ob1
Summary:
## Stack
Now that we've migrated `require` calls to `import` (with `interopRequireDefault`) we can safely change `module.exports` and `exports` assignments to ESM-syntax `export`s.

Note that changing a CommonJS export to an `export default` would be a breaking change, but all of the files changed in that way are `/private/`. All public entry files use named exports, which don't require an interop helper.

We do this with the help of `import/no-commonjs` ESLint rule, although the actual changes are largely manual + AI.

## This diff
Three more subpackages covering source mapping and friends.

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D79804097

fbshipit-source-id: 31f132fdf4937eda37eac6e6252cf527e649d90a