Convert to ESM-syntax exports: metro (3/3)
Summary:
## Stack
Now that we've migrated `require` calls to `import` (with `interopRequireDefault`) we can safely change `module.exports` and `exports` assignments to ESM-syntax `export`s.

Note that changing a CommonJS export to an `export default` would be a breaking change, but all of the files changed in that way are `/private/`. All public entry files use named exports, which don't require an interop helper.

We do this with the help of `import/no-commonjs` ESLint rule, although the actual changes are largely manual + AI.

## This diff
Completes migrating the main `metro` package.

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D79799084

fbshipit-source-id: a364e7cc862e84902ae7dd8b8a9d2d73aed4aefd