Add temporary lint/no-commonjs-require
Summary:
We want to convert Metro to ESM syntax in order to make better use of Flow's API translator (and for modernisation and consistency).

The first step will be to rewrite `require` calls as static `import`s. This adds a temporary linter to help us.

Why temporary? [`import/no-commonjs`](https://github.com/import-js/eslint-plugin-import/blob/main/docs/rules/no-commonjs.md) is the OSS standard for this purpose, but unlike the rule in this diff it has no auto-fix, which will be useful. `import/no-commonjs` also doesn't allow us to lint against `require` while allowing `module.exports`.

Why commit this at all? The rule is deleted at the end of this stack in favour of `import/commonjs`, but we'll need it throughout the stack for test plans. And, we may want to restore it later for other parts of the codebase.

Reviewed By: vzaidman

Differential Revision: D79816826

fbshipit-source-id: 52af2092404d5ffc211a2405ac20fe8260b0da67