Convert to ESM-syntax exports: metro (2/3)
Summary:
## Stack
Now that we've migrated `require` calls to `import` (with `interopRequireDefault`) we can safely change `module.exports` and `exports` assignments to ESM-syntax `export`s.

Note that changing a CommonJS export to an `export default` would be a breaking change, but all of the files changed in that way are `/private/`. All public entry files use named exports, which don't require an interop helper.

We do this with the help of `import/no-commonjs` ESLint rule, although the actual changes are largely manual + AI.

## This diff
Continues converting the `metro` package

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D79741120

fbshipit-source-id: 6474821d169fecb6735e55759344b4e870a6fecb