Convert to ESM-syntax exports: metro-file-map
Summary:
## Stack
Now that we've migrated `require` calls to `import` (with `interopRequireDefault`) we can safely change `module.exports` and `exports` assignments to ESM-syntax `export`s.

Note that changing a CommonJS export to an `export default` would be a breaking change, but all of the files changed in that way are `/private/`. All public entry files use named exports, which don't require an interop helper.

We do this with the help of `import/no-commonjs` ESLint rule, although the actual changes are largely manual + AI.

## This diff
Migrates `metro-file-map` (this one is particularly useful, because it lends itself to upstreaming as TypeScript back to `jest-haste-map`).

Note that `worker.js` and dependencies remain CommonJS so they can run directly from source.

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D79804093

fbshipit-source-id: f6c5dd800c4b348db858316cc7e410bbfcac7565