Convert to ESM-syntax exports: metro-transform-worker,metro-transform-plugins
Summary:
## Stack
Now that we've migrated `require` calls to `import` (with `interopRequireDefault`) we can safely change `module.exports` and `exports` assignments to ESM-syntax `export`s.

Note that changing a CommonJS export to an `export default` would be a breaking change, but all of the files changed in that way are `/private/`. All public entry files use named exports, which don't require an interop helper.

We do this with the help of `import/no-commonjs` ESLint rule, although the actual changes are largely manual + AI.

## This diff
Migrates two more packages.

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D79804096

fbshipit-source-id: dd3b6802c0a8667321f533ee140847514702f763