Convert to ESM-syntax exports: metro (1/3)
Summary:
## Stack
Now that we've migrated `require` calls to `import` (with `interopRequireDefault`) we can safely change `module.exports` and `exports` assignments to ESM-syntax `export`s.

Note that changing a CommonJS export to an `export default` would be a breaking change, but all of the files changed in that way are `/private/`. All public entry files use named exports, which don't require an interop helper.

We do this with the help of `import/no-commonjs` ESLint rule, although the actual changes are largely manual + AI.

## This diff
`metro` is pretty big, it turns out - this diff tackles part of the `metro` package.

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D79735956

fbshipit-source-id: 096bd50472701930befbc21f2483c09bacd55d99