Convert to ESM-syntax exports: buck-worker-tool,metro-minify-terser,metro-babel-*
Summary:
## Stack
Now that we've migrated `require` calls to `import` (with `interopRequireDefault`) we can safely change `module.exports` and `exports` assignments to ESM-syntax `export`s.

Note that changing a CommonJS export to an `export default` would be a breaking change, but all of the files changed in that way are `/private/`. All public entry files use named exports, which don't require an interop helper.

We do this with the help of `import/no-commonjs` ESLint rule, although the actual changes are largely manual + AI.

## This diff
Migrates four more subpackages

Changelog: Internal

Reviewed By: vzaidman

Differential Revision: D79804094

fbshipit-source-id: 0dfd04603cd7b259c42664ec6f47f7b8c54b9a1c