support non-ASCII filesystem bundle paths with metro server (#1538)
Summary:
In a codebase with non-ascii characters in file paths (e.g. Japanese or Chinese characters), metro has an error when serving bundles:

```
TypeError: Invalid character in header content
```

To fix it, we decode the URL received in `parseBundleOptionsFromBundleRequestUrl`.

Notice that this is just a quick fix to the bug, however hot reload and other parts of the system would still not work with non-ascii characters.

This will be fixed in the next diff.

Changelog: [Fixed] Initial metro support to non-ASCII bundle/source files/sourcemaps

Pull Request resolved: https://github.com/facebook/metro/pull/1538

Test Plan:
* `js1 test parseBundleOptionsFromBundleRequestUrl-test.js`
* Hot reload works the same with regular ascii filenames
* Also- using the test diff D78806803 I managed to reproduce Metro not launching RNTester before this fix:

1. `buck2 install -r rntester-android` (with the modified bundle path)
2. `DEV=1 js1 run`
3. launch `RNTester`

Before- app crashes:
 https://pxl.cl/7LFkb

After- app launches, debugging works:
{F1980565661}

Can even add the bundle to ignore-list and then it gets ignored:
{F1980565660}

Reviewed By: robhogan

Differential Revision: D78799358

Pulled By: vzaidman

fbshipit-source-id: 1d5b1541f594a1a64479a66947528a8ab85a4518